refactor(submission): update status enum and introduce EnrichmentService for evidence processing
- Replace SubmissionStatus values with PENDING_ANALYSIS, ENRICHED, ACCEPTED, and REJECTED
- Change default status to PENDING_ANALYSIS on submission
- Extract enrichment logic to new EnrichmentService
- Rename SubmissionEnrichmentListener to MarkEvidenceSubmittedListener and delegate enrichment
