+++ intake/src/main/java/pt/estga/intake/listeners/MarkEvidenceSubmittedListener.java
@@ -1,0 +1,27 @@
+package pt.estga.intake.listeners;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.scheduling.annotation.Async;
+import org.springframework.stereotype.Component;
+import org.springframework.transaction.event.TransactionPhase;
+import org.springframework.transaction.event.TransactionalEventListener;
+import pt.estga.intake.events.MarkEvidenceSubmittedEvent;
+import pt.estga.intake.services.EnrichmentService;
+
+@Component
+@RequiredArgsConstructor
+@Slf4j
+public class MarkEvidenceSubmittedListener {
+
+    private final EnrichmentService enrichmentService;
+
+    @Async
+    @TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT)
+    public void enrichMarkEvidence(MarkEvidenceSubmittedEvent event) {
+        Long submissionId = event.getSubmissionId();
+        log.info("Async detection processing for submission ID: {}", submissionId);
+
+        enrichmentService.enrichSubmission(submissionId);
+    }
+}
